Treat Quark floats as Double instances in Java

+++ quark/test/emit/simple-math/java/Functions.java
@@ -4,12 +4,12 @@ public class Functions {
         System.out.println(Integer.toString(four));
         Integer half = ~((~(1)) / (2));
         System.out.println(Integer.toString(half));
-        float num = (float) (314);
-        float den = (float) (100);
-        float pi = (num) / (den);
-        System.out.println(Float.toString(pi));
-        float pie = (float) (3.14);
-        System.out.println(Float.toString(pie));
+        Double num = (Double) (314.0);
+        Double den = (Double) (100.0);
+        Double pi = (num) / (den);
+        System.out.println(Double.toString(pi));
+        Double pie = (Double) (3.14);
+        System.out.println(Double.toString(pie));
         Integer n = ~((~(-(100))) / (3));
         System.out.println(Integer.toString(n));
         Integer m = ~((~(100)) / (-(3)));
